added blog of resolving vulnerabilities
