--- blog/2023-12-31-resolve-vulnerabilities-in-react-js.md
@@ -1,9 +1,0 @@
----
-slug: resolve-vulnerabilities-in-react-js
-title: Resolve vulnerabilities in React.js
-description: Resolve vulnerabilities in front-end libraries like React.js, Angular, and Vue.js
-authors: [utk09]
-tags: [vulnerabilities, react, angular, vue, front-end, security, libraries]
----
-
-If you are using React.js, Angular, Vue.js or if your framework uses ```package.json``` in your project, you might have come across vulnerabilities when you run `npm audit` or `yarn audit`. These vulnerabilities can be a security risk and should be resolved as soon as possible. In this article, we will discuss how to resolve vulnerabilities in React.js.
