+++ src/components/HomepageHeader/QuoteComponent.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const QuoteComponent = () => (
+  <p className="custom-mono text-xs my-6 text-center">
+    "If you get to a certain achievement/platform, it's your unsaid duty to help
+    others reach the same level in lesser time." ~Anonymous
+  </p>
+);
+
+export default QuoteComponent;
