+++ tutorials/tutorialsData.js
@@ -2,7 +2,7 @@ const tutorialsData = [
   {
     id: 1,
     title: "Data Science Part 1 - Getting the Data",
-    link: "/tutorials/data-science-part-1-getting-the-data",
+    link: "/tutorials/data-science/data-science-part-1-getting-the-data",
     description: "Learn how to get the data for your data science project.",
     status: "In Progress",
   },
